requirements.txt
Pillow==11.1.0
click==8.1.8
dearpygui==2.0.0
imagequant==1.1.3
mozjpeg_lossless_optimization==1.1.5
numpy==2.2.2
zopflipy==1.11
